Refactor: VolumeControl에는 볼륨 관련 코드만 포함되게! -> PlayerManager 파일로 handleAutoPlayMusic 이동
