+++ src/main/java/discord/podongbot/response/ChattingReaction.java
@@ -1,7 +1,6 @@
 package discord.podongbot.response;
 
 import discord.podongbot.music.PlayerManager;
-import discord.podongbot.volume.VolumeControl;
 import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
@@ -14,7 +13,7 @@ public class ChattingReaction extends ListenerAdapter {
         String msg = event.getMessage().getContentRaw();
         String[] parts = msg.split(" ", 2);
 
-        VolumeControl.handleAutoPlayMusic(event); // 음악 자동 재생
+        PlayerManager.handleAutoPlayMusic(event); // 음악 자동 재생
 
         switch(parts[0]) {
             case "!노래":
